clean up worker memory management
